Incorporate some fixes from snes9xgx Less freetype texture size

+++ source/menu.cpp
@@ -206,7 +206,7 @@ UpdateGUI (void *arg)
 				for(i = 0; i <= 255; i += 15)
 				{
 					mainWindow->Draw();
-					Menu_DrawRectangle(0,0,screenwidth,screenheight,(GXColor){0, 0, 0, i},1);
+					Menu_DrawRectangle(0,0,screenwidth,screenheight,(GXColor){0, 0, 0, (u8)i},1);
 					Menu_Render();
 				}
 				ExitApp();
